refactor: remove init data in production
