+++ test/api_rest_test/internal/util.go
@@ -56,6 +56,7 @@ func CreateServerRouterForApiTest() *serverRouter {
 	userRepo := memdb.NewUserRepo()
 	todoRepo := memdb.NewTodoRepo()
 	initUserData(userRepo)
+	initTodoData(todoRepo)
 
 	server := rest.NewServer(userRepo, todoRepo)
 	return &serverRouter{
@@ -64,7 +65,6 @@ func CreateServerRouterForApiTest() *serverRouter {
 		todoRepo: todoRepo,
 	}
 }
-
 func initUserData(userRepo repo.UserRepo) {
 	dao := entity.User{
 		Username:  "daopham",
@@ -80,6 +80,28 @@ func initUserData(userRepo repo.UserRepo) {
 	hung.SetPassword("87654321")
 	userRepo.SaveUser(&dao)
 	userRepo.SaveUser(&hung)
+}
+
+func initTodoData(todoRepo repo.TodoRepo) {
+	daoTodo := entity.Todo{
+		Username: "daopham",
+		Content:  "Dao walking",
+		IsDone:   false,
+	}
+	daoTodo2 := entity.Todo{
+		Username: "daopham",
+		Content:  "Dao learning",
+		IsDone:   false,
+	}
+	hungTodo := entity.Todo{
+		Username: "hungpham",
+		Content:  "Hung working",
+		IsDone:   false,
+	}
+
+	todoRepo.CreateTodo(&daoTodo)
+	todoRepo.CreateTodo(&daoTodo2)
+	todoRepo.CreateTodo(&hungTodo)
 }
 
 func CheckIfTodoListContainsTodo(todos []entity.Todo, target entity.Todo) bool {
